+++ examples/run_prototype.jl
@@ -12,20 +12,24 @@
 ##########################################################################
 #using EnergyModelsGUI
 
-
-#1: generate a CE model that is loaded into memory as "case".
-#include("generate_EMG.jl")
-#include("generate_EMB.jl")
-include("generate_EMI.jl")
-
-#2: load the functions and data definitions:
+#1: load the functions and data definitions:
 include(joinpath(@__DIR__,"..","src", "EnergyModelsGUI.jl"))
+
+idToColorsMap::Dict{Any,Any} = Dict{Any, Any}() # Initialize dictionary for colors map
+idToIconsMap::Dict{Any,Any} = Dict{Any, Any}() # Initialize dictionary for icons map
+
+
+#2: generate a CE model that is loaded into memory as "case".
+#include("generate_EMB.jl")
+#include("generate_EMB_sink_source.jl")
+include("generate_EMG.jl")
+#include("generate_EMI.jl")
 
 #3: Specify path for design information:
 path = joinpath(@__DIR__, "..", "design") # folder where visualization info is saved and retrieved
 
 #4: Generate the system topology:
-design = EnergyModelsGUI.EnergySystemDesign(case, path);
+design = EnergyModelsGUI.EnergySystemDesign(case, path; idToColorsMap, idToIconsMap)
 
 #5: Plot the topology:
 EnergyModelsGUI.view(design)