+++ src/EnergyModelsGUI.jl
@@ -10,15 +10,18 @@ using TOML
 using FilterHelpers
 using Observables
 using EnergyModelsBase
+using EnergyModelsGeography
 using Colors
 using GeoMakie, GeoJSON
 using GLMakie, GeometryBasics
 using CairoMakie
 
+missingColor = :black # Default color when color is not provided
 include("structureTopology.jl")
 include("viewTopology.jl")
 
 
 export EnergySystemDesign
+export setColors!
 
 end # module